Converts Build Script to PowerShell (!41)
In a push to automate project builds, this changes the build script from a BAT script to a PowerShell script. This is a cross-platform solution that should help us build the project within the mcr.microsoft.com/dotnet/sdk:6.0 docker image (Debian).

Also updates the README to list PowerShell 7 as a dependency as the pwsh tool is not available in Windows PowerShell (v5).

Co-authored-by: Tyler Brownell <brownelltyler@gmail.com>
Reviewed-on: https://dev.sp-tarkov.com/SPT-AKI/Launcher/pulls/41
Co-authored-by: Refringe <refringe@noreply.dev.sp-tarkov.com>
Co-committed-by: Refringe <refringe@noreply.dev.sp-tarkov.com>